fixing again

+++ bin/binarize.sh
@@ -134,8 +134,10 @@ do
       python_files_in_current_dir=`ls -1 *.py 2>/dev/null | wc -l`
       if [ $python_files_in_current_dir -eq 0 ]
       then
-         mv ./tmp_init_file_dir/__init__.py .
-         rm -rf ./tmp_init_file_dir
+         if [ -d "tmp_init_file_dir" ]; then
+            mv ./tmp_init_file_dir/__init__.py .
+            rm -rf ./tmp_init_file_dir
+         fi
          cd ..
          continue
       fi
